move/rename extras file for clarity
